Moved SnowPistons to package src.com.github.richardyin.jeevesCraft
Figured it was appropriate
