+++ src/com/github/abujaki21/jeevesCraft/JeevesCraft.java
@@ -20,11 +20,14 @@ package com.github.abujaki21.jeevesCraft;
 import java.io.File;
 import java.io.IOException;
 import java.util.logging.Logger;
+
 import org.bukkit.configuration.file.YamlConfiguration;
 import org.bukkit.configuration.InvalidConfigurationException;
 import org.bukkit.event.Listener;
 import org.bukkit.plugin.java.JavaPlugin;
 import org.bukkit.Server;
+
+import com.github.richardyin.jeevesCraft.SnowPistons;
 
 public final class JeevesCraft extends JavaPlugin implements Listener{
 	private Logger logger;
